Intake timeout, auton skills, and rumble changes
Added a function to let the intake timeout if it gets stuck during auton, and updated autons to use it. Changed auton skills and added a bunch of comments to it. Made it stop rumbling controller at the end of the match even if expansion was not fired. Also added a function to shoot multiple times, and updated autons to use it. I forgot to commit for a while so there's probably also other stuff in there that I missed, but that's most of it.